Add jet roll, separate output from simulation
By updating the output at the end, we can simplify the simulation logic. Because we only ever set things to unlocked (we never set cruiser_unlocked back to false), we don't have to worry about continuing to roll when it's already been unlocked. The worst that would happen is it's set to true again when it's already true.